pyproject.toml
[project]
name = "raketa"
version = "0.1.0"
description = "no description"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika==9.5.5",
    "aiofiles==24.1.0",
    "aiogram==3.17.0",
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiormq==6.8.1",
    "aiosignal==1.3.2",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asyncpg==0.30.0",
    "attrs==25.1.0",
    "billiard==4.2.1",
    "branca==0.8.1",
    "cachetools==5.5.2",
    "celery==5.5.1",
    "certifi==2025.1.31",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "distro==1.9.0",
    "dramatiq==1.17.1",
    "emoji==2.14.1",
    "exceptiongroup==1.2.2",
    "frozenlist==1.5.0",
    "fuzzywuzzy==0.18.0",
    "geographiclib==2.0",
    "geopy==2.4.1",
    "gevent==24.11.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.2",
    "google-api-python-client==2.167.0",
    "google-auth==2.39.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.1.1",
    "grpcio==1.72.0rc1",
    "grpcio-status==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.0.0",
    "Jinja2==3.1.5",
    "jiter==0.8.2",
    "kombu==5.5.2",
    "Levenshtein==0.26.1",
    "lxml==5.3.0",
    "magic-filter==1.0.12",
    "MarkupSafe==3.0.2",
    "multidict==6.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.2.3",
    "openai==1.44.1",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pamqp==3.3.0",
    "parselab==0.1.8",
    "pillow==11.1.0",
    "pluggy==1.5.0",
    "prometheus_client==0.21.1",
    "prompt_toolkit==3.0.50",
    "propcache==0.2.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "psycopg2==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.10.6",
    "pydantic-settings==2.5.0",
    "pydantic_core==2.27.2",
    "pydub==0.25.1",
    "PyMuPDF==1.25.4",
    "pyparsing==3.2.3",
    "PySocks==1.7.1",
    "pytest==8.3.3",
    "pytest-asyncio==0.25.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-Levenshtein==0.26.1",
    "pytz==2024.2",
    "RapidFuzz==3.12.1",
    "redis==5.2.1",
    "reportlab==4.3.1",
    "requests==2.32.3",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "SpeechRecognition==3.10.4",
    "SQLAlchemy==2.0.34",
    "tqdm==4.67.1",
    "trio==0.29.0",
    "trio-websocket==0.12.2",
    "typing_extensions==4.12.2",
    "tzdata==2025.2",
    "uritemplate==4.1.1",
    "urllib3==2.2.2",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
    "xyzservices==2025.1.0",
    "yarl==1.18.3",
    "zope.event==5.0",
    "zope.interface==7.2",
    "redis==5.2.1",
    "apscheduler>=3.11.0",
]
